pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-spec-driven-pipeline"
version = "0.1.0"
description = "Specification-driven agent to generate data pipeline code (Dataform, DBT, Spark) and tests from requirements"
readme = "README.md"
authors = [{ name = "Lead Data Engineer" }]
requires-python = ">=3.9"

# Core runtime dependencies
# Keep minimal to avoid conflicts; BigQuery deps and LLM fallback are optional extras
dependencies = [
  "google-adk>=0.2.0",
  "typer>=0.12.5",
  "PyYAML>=6.0.2",
  "sqlparse>=0.5.1",
  "faker>=26.0.0",
  "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
# Install with: pip install -e .[bigquery]
bigquery = [
  "google-cloud-bigquery>=3.25.0",
  "pandas>=2.2.2",
  "pyarrow>=17.0.0",
]
# Developer/test tools: pip install -e .[dev]
dev = [
  "pytest>=8.3.3",
]
# LLM fallback client for direct Gemini usage: pip install -e .[llm]
llm = [
  "google-generativeai>=0.8.3",
]
# Web API/UI: pip install -e .[web]
web = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
